requirements.txt
GitPython==2.1.5
boto3==1.4.6
requests==2.18.4
